added initAll route dbname and initData env vars
