+++ server/startup/db.js
@@ -1,11 +1,12 @@
 const mongoose = require('mongoose');
 const { logger } = require('../middleware/winston'); // middleware/error.js which is running [npm] winston for error handling
+const config = require('config');
 
 // DB Config | Saved in the config folder | Mogoose Key
 const dbURI = require('../config/keys').mongoURI;
 const mongoOptions =  {
     useNewUrlParser: true,
-    dbName: 'test',
+    dbName: config.get('dbName'),
     useFindAndModify: false,
     useCreateIndex: true,
     useUnifiedTopology: true
