+++ server/startup/routes.js
@@ -15,6 +15,8 @@ const users = require('../routes/users');
 const news = require('../routes/api/news');
 const logs = require('../routes/api/log');
 const banking = require('../routes/api/banking');
+const initData = require('../routes/api/initData');
+
 
 module.exports = function(app) {
     // Bodyparser Middleware
@@ -34,6 +36,7 @@ module.exports = function(app) {
     app.use('/api/news', news); // Route for the news desks
     app.use('/api/logs', logs); // Route for logs
     app.use('/api/banking', banking); // Route for banking functions
+    app.use('/api/initData', initData); // Route for Init Data functions
 
     app.use(error.routeError);
 }
