+++ server/util/baseLoad.js
@@ -1,5 +1,6 @@
-const fs = require('fs')
-const file = fs.readFileSync('./init-json/initbase.json', 'utf8');
+const fs = require('fs');
+const config = require('config');
+const file = fs.readFileSync(config.get('initPath') + 'init-json/initbase.json', 'utf8');
 const baseDataIn = JSON.parse(file);
 //const mongoose = require('mongoose');
 const baseLoadDebugger = require('debug')('app:baseLoad');
