+++ server/util/config/custom-environment-variables.json
@@ -1,5 +1,4 @@
 {
-    "jwtPrivateKey": "jwtPrivateKey",
     "initPath": "initDataDir",
     "dbName": "dbName"
 }
