+++ server/util/config/default.json
@@ -1,0 +1,4 @@
+{
+    "initPath": "./",
+    "dbName": "rawr"
+}
