+++ server/util/initRefLoad.js
@@ -1,5 +1,6 @@
-const fs = require('fs')
-const file = fs.readFileSync('./init-json/refdata.json', 'utf8');
+const fs = require('fs');
+const config = require('config');
+const file = fs.readFileSync(config.get('initPath') + 'init-json/refdata.json', 'utf8');
 const refDataIn = JSON.parse(file);
 //const mongoose = require('mongoose');
 const zoneInitDebugger = require('debug')('app:zoneInit');
